Update the way to enable events
Currently, the way to enable events is to use predefined arguments.
They are only compatible with some events, and may not be supported
by the device.
This updates pmugraph to use device's perf event arguments.
With that change, any event can be enabled individually.
And, this will always works, whatever is the device.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>